Show and fit histograms
show_hist.C is updated to read from new ROOT file with the TH2F
histograms and its new names.

show_prof.C is added to show the TProfile histograms from its own ROOT
file.

show_fit.C is added to show the fit of TProfiles and the betta function.

getFit.C is used as an alternate way to get the fit. It reads the data
from the particle_data.root file (which has a TTree inside), creates
the histogram (for the particle given as parameter) and then fits it.